+++ __tests__/utils/commands.test.js
@@ -1,7 +1,6 @@
 const rewire = require('rewire');
-const { initCommands } = require('../commands');
 
-const DEFAULT_CONFIG = rewire('../config').__get__('DEFAULT_CONFIG');
+const DEFAULT_CONFIG = rewire('../../config').__get__('DEFAULT_CONFIG');
 
 global.Cypress = {
   env: () => {},
@@ -13,21 +12,7 @@ global.Cypress = {
 
 global.cy = {};
 
-describe('command', () => {
-  it('should create command', () => {
-    global.before = jest.fn();
-    global.after = jest.fn();
-
-    initCommands();
-    global.cy.task = jest.fn().mockResolvedValue({ pass: true });
-
-    expect(global.Cypress.Commands.add).toBeCalled();
-    expect(global.Cypress.Commands.add.mock.calls.length).toEqual(1);
-    expect(global.Cypress.Commands.add.mock.calls[0][0]).toEqual('toMatchSnapshot');
-    expect(global.after).toBeCalled();
-    expect(global.before).toBeCalled();
-  });
-
+describe('utils/command', () => {
   describe('should retrieve config', () => {
     const CONFIG = {
       foo: 'bar',
@@ -42,7 +27,7 @@ describe('command', () => {
         return returnValue;
       };
 
-      const commands = rewire('../commands');
+      const commands = rewire('../../utils/commands');
       commands.__set__('Cypress', global.Cypress);
 
       const getConfig = commands.__get__('getConfig');
@@ -52,7 +37,7 @@ describe('command', () => {
 
     it('with config', () => {
       global.Cypress.env = () => CONFIG;
-      const commands = rewire('../commands');
+      const commands = rewire('../../utils/commands');
       commands.__set__('Cypress', global.Cypress);
 
       const getConfig = commands.__get__('getConfig');
@@ -61,7 +46,7 @@ describe('command', () => {
 
     it('without config - error should be thrown', () => {
       global.Cypress.env = () => undefined;
-      const commands = rewire('../commands');
+      const commands = rewire('../../utils/commands');
       commands.__set__('Cypress', global.Cypress);
 
       const getConfig = commands.__get__('getConfig');