+++ commands.js
@@ -4,74 +4,66 @@ const { merge, cloneDeep } = require('lodash');
 const { MATCH } = require('./tasks/task-names');
 const { initUi } = require('./ui');
 const {
-  formatNormalizedJson, getTestTitle, getSnapshotTitle, getSnapshotFilename, snapshotTitleIsUsed,
+  getTestTitle,
+  getSnapshotTitle,
 } = require('./utils/snapshots');
-const { CONFIG_KEY } = require('./config');
+const {
+  cleanUpSnapshots,
+  fixConfig,
+  getConfig,
+  getTestForTask,
+  getSubject,
+  isHtml,
+} = require('./utils/commands');
+const {
+  TYPE_JSON,
+  TYPE_HTML
+} = require('./constants');
 
 const URL_PREFIX = '#cypress-plugin-snapshot-';
 const COMMAND_NAME = 'toMatchSnapshot';
 const NO_LOG = { log: false };
 
-function getConfig() {
-  fixConfig();
-  const config = Cypress.env(CONFIG_KEY);
-  if (!config) {
-    throw new Error('Config cannot be found.');
-  }
+function commandToMatchSnapshot(testSubject, options) {
+  const test = getTestForTask();
+  const testTitle = getTestTitle(test);
+  const testFile = Cypress.spec.absolute;
+  const snapshotTitle = getSnapshotTitle(test);
 
-  return config;
-}
+  const subject = getSubject(testSubject);
+  const dataType = isHtml(testSubject) ? TYPE_HTML : TYPE_JSON;
 
-// Removes unused snapshots from snapshot file
-function cleanUpSnapshots() {
-  const config = getConfig();
-  if (!config.autoCleanUp) {
-    return;
-  }
+  const toMatchSnapshot = (result) => {
+    const args = window.parent.window.btoa(JSON.stringify(result));
+    const passedMessage = result.expected ? 'Snapshots match' : 'Snapshot created, autopassed';
+    const message = result.passed ?
+        passedMessage
+      : `[compare snapshot](${URL_PREFIX}${args})`;
 
-  const filename = getSnapshotFilename(Cypress.spec.relative);
-  cy.readFile(filename, NO_LOG).then((content) => {
-    if (content) {
-      const snapshot = JSON.parse(content);
-      const keys = Object.keys(snapshot);
+    const log = Cypress.log({
+      $el: subject,
+      name: COMMAND_NAME,
+      displayName: 'snapshot',
+      message,
+      consoleProps: () => result,
+    });
 
-      const cleanSnapshot = keys
-        .filter(snapshotTitleIsUsed)
-        .reduce((result, key) => {
-          result[key] = snapshot[key];
-          return result;
-        }, {});
-
-      cy.writeFile(filename,
-        formatNormalizedJson(cleanSnapshot),
-        NO_LOG);
+    if (!result.passed) {
+      log.set('state', 'failed');
+      throw new Error(`Snapshots do not match:\n${result.diff}`);
     }
-  });
-}
 
-function getTest() {
-  return Cypress.mocha.getRunner().test;
-}
-
-function getTestForTask(test) {
-  if (!test) {
-    test = getTest();
-  }
-  return {
-    id: test.id,
-    title: test.title,
-    parent: test.parent && test.parent.title ? getTestForTask(test.parent) : null,
+    return subject;
   };
-}
 
-/**
- * Check if config in `Cypress.env` is stringified JSON.
- * If so parse it and set the parsed value back in `Cypress.env`.
- */
-function fixConfig() {
-  if (typeof Cypress.env(CONFIG_KEY) === 'string') {
-    Cypress.env(CONFIG_KEY, JSON.parse(Cypress.env(CONFIG_KEY)));
-  }
+  return cy.task(MATCH, {
+    dataType,
+    options,
+    snapshotTitle,
+    subject,
+    testFile,
+    testTitle,
+  }, NO_LOG).then(toMatchSnapshot);
 }
 
 function initCommands() {
@@ -87,47 +79,14 @@ function initCommands() {
     cleanUpSnapshots();
   });
 
-  Cypress.Commands.add(COMMAND_NAME, { prevSubject: 'optional' }, (subject, taskOptions) => {
-    if (!subject) {
-      return subject;
+  Cypress.Commands.add(COMMAND_NAME, { prevSubject: 'optional' }, (commandSubject, taskOptions) => {
+    if (!commandSubject) {
+      return commandSubject;
     }
 
     const options = merge({}, cloneDeep(getConfig()), taskOptions);
-
-    const test = getTestForTask();
-    const testTitle = getTestTitle(test);
-    const snapshotTitle = getSnapshotTitle(test);
-
-    const toMatchSnapshot = (result) => {
-      const args = window.parent.window.btoa(JSON.stringify(result));
-      const passedMessage = result.expected ? 'Snapshots match' : 'Snapshot created, autopassed';
-      const message = result.passed
-        ? passedMessage
-        : `[compare snapshot](${URL_PREFIX}${args})`;
-
-      const log = Cypress.log({
-        $el: subject,
-        name: COMMAND_NAME,
-        displayName: 'snapshot',
-        message,
-        consoleProps: () => result,
-      });
-
-      if (!result.passed) {
-        log.set('state', 'failed');
-        throw new Error(`Snapshots do not match:\n${result.diff}`);
-      }
-
-      return subject;
-    };
-
-    return cy.task(MATCH, {
-      testTitle,
-      testFile: Cypress.spec.absolute,
-      snapshotTitle,
-      subject,
-      options,
-    }, NO_LOG).then(toMatchSnapshot);
+    return cy.wrap(commandSubject)
+      .then((subject) => commandToMatchSnapshot(subject, options));
   });
 }