+++ constants.js
@@ -1,0 +1,4 @@
+module.exports = {
+  TYPE_HTML: 'html',
+  TYPE_JSON: 'json',
+}