+++ CHANGELOG.md
@@ -1,10 +1,13 @@
-1.0.7 / 2018-11-07
+1.1.0 / 2018-11-07
 ==================
 
+  * Change format of `.snap` files to have a nicer format for diffing HTML in `git`
   * Make sure config is always available (fixes #2)
   * Resolve dependency paths relative to plugin location (fixes #3)
   * Moved CSS to `assets/styles.css`
   * Moved javascript to `assets/script.js`
+  * Added support for DOM elements
+  * Added `prettier` for formatting HTML before comparing
 
 1.0.6 / 2018-11-07
 ==================